ocr-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
Pillow==10.4.0
pytesseract==0.3.10
python-multipart==0.0.9

